CR: use-file-upload-hook
